Rename ClassificationEditor to CategoryEditor
