+++ frontend_react/src/components/CategoryEditor/CategoryGroup.jsx
