+++ frontend_react/src/components/CategoryEditor/EditTabSwitcher.jsx
