+++ frontend_react/src/components/CategoryEditor/GameTagEditor.jsx
