+++ frontend_react/src/components/CategoryEditor/GameTagsGroup.jsx
